+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+# 0.0.2
+
+## Features
+
+- Create default config file in current floder
+- Add config example for user
+- Finish all the componet
+
 # 0.0.1
 
 ## Features
 
-- Allow user to create resource in one commandgit 
+- Allow user to create resource in one command